Move job report to another groovy file so that it can be used
Also change to only printing the output to the console because we cannot print to files due to security restrictions
