+++ jobs/generation/MetaGenerator.groovy
@@ -254,6 +254,8 @@ repos.each { repoInfo ->
             
             // 5 second quiet period before the job can be scheduled
             quietPeriod(5)
+            
+            timestamps()
         }
         
         if (isPRTest) {
